chore: fixing post req
